Removed verbose error messages, added missing init failure check. Fix in rc.sensors, uncommented hmc5883 driver start.
Moved ioctl handlers to CDev-inheriting classes to get clean escalation of unhandled ioctls to superclass
